pyproject.toml
[project]
name = "ftmo-academy"
version = "0.1.0"
description = "FTMO Academy content optimization and multi-model testing"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0",
    "google-genai>=1.0",
    "anthropic>=0.40",
    "python-dotenv>=1.0",
    "pandas>=2.0",
    "openpyxl>=3.1",
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "html2text>=2024.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8",
    "mypy>=1.13",
    "pytest>=8.0",
    "pandas-stubs>=2.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
